requirements.txt
gradio>=3.50.2
python-dotenv>=1.0.0
aiohttp>=3.8.5
Pillow>=10.0.0
asyncio>=3.4.3
requests>=2.31.0
tqdm>=4.65.0
numpy>=1.24.3
python-dotenv>=1.0.0
huggingface-hub>=0.16.4
transformers>=4.31.0
accelerate>=0.21.0
safetensors>=0.3.1
